Refactor the code split Comments to 2 files.
