+++ src/style.css
@@ -17,7 +17,7 @@ footer {
 
 .comments-form {
   text-align: left;
-  width: 50%;
+  width: 75%;
   margin: auto;
 }
