+++ src/utils/populatePopup.js
@@ -1,29 +1,4 @@
-const appId = 'OJhoS4niRmFdRpqldNlB';
-const commentsURL = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}/comments`;
-
-export const addComment = async (comment) => {
-  await fetch(commentsURL, {
-    method: 'POST',
-    body: JSON.stringify(comment),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  });
-};
-
-export const getComments = async (i) => {
-  const commentsContainer = document.getElementById(`comments${i}`);
-  commentsContainer.innerHTML = '';
-  await fetch(`${commentsURL}?item_id=item${i}`)
-    .then((response) => response.json())
-    .then((comments) => {
-      comments.forEach((comment) => {
-        const commentText = document.createElement('p');
-        commentText.innerHTML = `${comment.creation_date} ${comment.username}: ${comment.comment}`;
-        commentsContainer.appendChild(commentText);
-      });
-    });
-};
+import {addComment, getComments} from './commentsHandler';
 
 const comments = (pokemon, commentLink, i) => {
   const container = document.getElementById('container');
@@ -41,18 +16,18 @@ const comments = (pokemon, commentLink, i) => {
     abilities.push(ability.ability.name);
   });
   commentPopup.innerHTML = `<div class="modal-dialog modal-xl">
-                              <div class="modal-content">
+                              <div class="modal-content container-fluid p-5">
                                 <div class="modal-header border-0">
                                   <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                                 </div>
                                 <div class="modal-body text-center">
-                                  <div class="">
+                                  <div class="p-5">
                                     <img src="${pokemon.sprites.other['official-artwork'].front_default}" width="200rem" alt="${pokemon.forms[0].name}" />
                                     <h3 class="modal-title mb-2" id="exampleModalLabel">${pokemon.forms[0].name.toUpperCase()}</h3>   
                                   </div>
-                                  <div class="d-flex justify-content-center">
-                                    <div class="row">
-                                    <div class="col-8  align-left characteristics">
+                                  <div class="p-5 ">
+                                    <div class="row p-5">
+                                    <div class="col-8  align-left characteristics ml-4">
                                         <p><b>Types:</b> ${types.join(', ')}</p>
                                         <p><b>Abilities:</b> ${abilities.join(', ')}</p>
                                       </div>
@@ -60,19 +35,16 @@ const comments = (pokemon, commentLink, i) => {
                                       <div class="col-4 align-left characteristics">
                                         <p><b>Weight:</b> ${pokemon.weight}kg</p>
                                         <p><b>Height:</b> ${pokemon.height}m</p>
-                                      </div>
-                                    
-                                      
-                                 
+                                      </div>    
                                     </div>
                                   </div>
-                                  <div class="d-flex flex-column justify-content-center">
+                                  <div class="">
                                     <h4>
                                         Comments (2)
                                     </h4>
-                                    <div id="comments${i}" class="align-left w-50 comments"></div>
+                                    <div id="comments${i}" class="align-left w-75 comments"></div>
                                   </div>
-                                  <div class="my-4 w-70">
+                                  <div class="my-4 ">
                                     <h4>
                                       Add a comment
                                     </h4>
